% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_rainfall.R
\name{max_rainfall}
\alias{max_rainfall}
\title{max_rainfall}
\usage{
max_rainfall(rain_data, location = "Brisbane", years = 2009:2017,
  month_spec = 4)
}
\arguments{
\item{rain_data}{data frame with columns Date, Location, Rainfall}

\item{location}{specified location in the data frame you want to see, default is Brisbane}

\item{years}{specified years in the data frame you want to see, default is 2009 to 2017}

\item{month_spec}{specified month in the data frame you want to see, default is 4 (april)}
}
\value{
Returns a list containing,
\describe{
 \item{Location}{Location specified}
 \item{Month}{Month specified}
 \item{Max_rainfall}{The maximum rainfall (mm) for each year in the dataset}
 }
}
\description{
Compute maximum precipitation each year for a specified location during a specified month and output that year, location, month, and precipitation level.
}
\author{
Gage Clawson
}
