% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_rainfall.R
\name{mean_rainfall}
\alias{mean_rainfall}
\title{mean_rainfall}
\usage{
mean_rainfall(rain_data, location = "Brisbane")
}
\arguments{
\item{rain_data}{data frame with columns Date, Location, and Rainfall}

\item{location}{specified location in the data frame you want to see, default is Brisbane}
}
\value{
Returns a list containing,
\describe{
 \item{Year}{Each year within the dataset}
 \item{Location}{Location specified}
 \item{Mean_rainfall}{The mean rainfall (mm) for each year in the dataset}
 }
}
\description{
Compute mean precipitation for all years in the dataset and a specified location and output that year, location, and precipitation level.
}
\author{
Gage Clawson
}
