% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_risks.R
\name{temp_risks}
\alias{temp_risks}
\title{temp_risks}
\usage{
temp_risks(data)
}
\arguments{
\item{data}{data frame with columns Date, Location, Temp3pm}
}
\value{
Returns a list containing,
\describe{
 \item{Location}{Location in Australia}
 \item{heat_stroke_n}{Number of days for a particular location where there has been a risk of heat stroke}
 \item{comfort_n}{Number of days for a particular location where the weather has been comfortable}
 \item{freezing_n}{Number of days for a particular location where there has been a risk of freezing}
 }
}
\description{
Compute the number of days per each location over the span of the data set where there is risk of heat stroke, comfortable weather, and freezing at 3 PM.
}
\author{
Gage Clawson
}
