% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_costs.R
\name{temperature_costs}
\alias{temperature_costs}
\title{temperature_costs}
\usage{
temperature_costs(data)
}
\arguments{
\item{data}{data frame with columns Date, MaxTemp}
}
\value{
Returns a list containing,
\describe{
 \item{Year}{Each year within the dataset}
 \item{hospital_bill}{Sum of potential hospital bills from heat stroke ($)}
 }
}
\description{
Compute the potential yearly hospital bill costs for people who might have a heat stroke because of the maximum temperature for that day over the span of the data set.
}
\author{
Gage Clawson
}
