% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtle_details.R
\name{turtle_details}
\alias{turtle_details}
\title{Estimated Turtle Age and Growth Rate}
\usage{
turtle_details(shell_length, maturity_length, growth_rate)
}
\arguments{
\item{maturity_length}{(yrs)}

\item{growth_rate}{(cm/yr)}

\item{shell_lenth}{(cm)}
}
\value{
estimated age (yrs) and growth rate (cm/yr)
}
\description{
This function determines the estimated age and growth rate of a sea turtle based
on a given shell length, maturity length, and growth rate
}
