% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_supply.R
\name{water_supply}
\alias{water_supply}
\title{water_supply}
\usage{
water_supply(rain_data, location = "Brisbane")
}
\arguments{
\item{rain_data}{data frame with columns Date, Rainfall}

\item{location}{specified location in the data frame you want to see, default is Brisbane}
}
\value{
Returns a list containing,
\describe{
 \item{Year}{Each year within the dataset}
 \item{Location}{Location specified}
 \item{annual_rainfall}{Sum of the rainfall for a year}
 \item{Water_supply_level}{Whether or not the water supply is good, okay, or bad}
 }
}
\description{
A list that outputs for each year, for a specified location, whether the water supply is good, okay, or bad.
}
\author{
Gage Clawson
}
