% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheat_temperature.R
\name{wheat_temperature}
\alias{wheat_temperature}
\title{wheat_temperature}
\usage{
wheat_temperature(
  daily_temp,
  age,
  risk_threshold_med = 75,
  risk_threshold_high = 82,
  age_threshold = 8
)
}
\arguments{
\item{daily_temp}{(C)}

\item{age}{(months)}

\item{risk_threshold_med}{(default = 75)}

\item{risk_threshold_high}{(default = 82)}

\item{age_threshold}{(years) (default = 8)}
}
\value{
risk (high, med, low), n_extremes
}
\description{
Compute level of risk associated with extreme temperature events for wheat crop
}
